# monay-caas/monay-enterprise-wallet/openapi.yaml
openapi: 3.0.3
info:
  title: Monay Enterprise Wallet API
  description: API documentation for Monay Enterprise Wallet - Token Management & Treasury Operations
  version: 1.0.0
  contact:
    name: Monay Support
    email: support@monay.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.monay.com/api
    description: Production server
...
paths:
...
  /wallets/{id}:
    get:
      tags: [Wallets]
      summary: Get wallet details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
...
  /billing/invoice/{month}:
    get:
      tags: [Billing]
      summary: Download invoice
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: month
          required: true
          schema:
            type: string
            pattern: '^\\d{4}-\\d{2}$'
      responses:
        '200':
          description: Invoice PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Wallet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [enterprise, treasury, consumer]
        address:
          type: string
        network:
          type: string
          enum: [base, solana, polygon]
        balance:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [active, frozen, closed]
        created_at:
          type: string
          format: date-time
...